backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
python-multipart==0.0.9
python-dotenv==1.0.1
supabase==2.9.1
anthropic==0.34.2
openpyxl==3.1.5
pydantic==2.9.2
aiofiles==24.1.0
httpx==0.27.2
